2010-07-13 03:03 UTC - kode54 - Implemented better end silence detection - Version is now 2.0.17
2010-07-11 03:49 UTC - kode54
- Updated PSX response filter to use all float storage

git-tfs-id: [http://localhost:8080/tfs/DefaultCollection/]$/foobar2000/files/plugins.root;C441

+++ plugins/foobar/foobar2000/foo_psf/PSXFilter.h
@@ -10,10 +10,9 @@ public:
 	virtual ~CPSXFilter();
 
 private:
-	int lx1,lx2,ly1,ly2;
-  float la0,la1,la2,lb1,lb2;
+	float lx1,lx2,ly1,ly2;
+	float la0,la1,la2,lb1,lb2;
 
-	int hx1[2],hx2[2],hy1[2],hy2[2];
-  float ha0,ha1,ha2,hb1,hb2;
-
+	float hx1[2],hx2[2],hy1[2],hy2[2];
+	float ha0,ha1,ha2,hb1,hb2;
 };
